backend/requirements.txt
fastapi
uvicorn
pydantic
pytest
requests
passlib[bcrypt]
psycopg2-binary
python-jose
dotenv
sqlalchemy
alembic
python-multipart
httpx
argon2-cffi
pytest-mock
jwt
email-validator
imbalanced-learn==0.12.4
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.5.2
joblib==1.4.2
